# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/location/": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve all locations from the database",
        "description": "This endpoint fetches a list of all locations in the database.",
        "responses": {
          "200": {
            "description": "List of all locations.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "streetAddress": {
                    "type": "string",
                    "example": "123 Main St"
                  },
                  "secondaryAddress": {
                    "type": "string",
                    "example": "Apt 4B"
                  },
                  "city": {
                    "type": "string",
                    "example": "Springfield"
                  },
                  "region": {
                    "type": "string",
                    "example": "Illinois"
                  },
                  "zipCode": {
                    "type": "string",
                    "example": "62701"
                  },
                  "country": {
                    "type": "string",
                    "example": "USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/stringify": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve all locations as stringified addresses",
        "description": "This endpoint fetches a list of all locations and returns them as formatted addresses.",
        "responses": {
          "200": {
            "description": "List of stringified addresses.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "123 Main St, Apt 4B, Springfield, Illinois, 62701, USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/stringify/city/{city}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve stringified locations filtered by city",
        "description": "This endpoint fetches all locations in a given city and returns them as formatted addresses.",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "City name to filter locations"
          }
        ],
        "responses": {
          "200": {
            "description": "List of stringified addresses for the specified city.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "123 Main St, Springfield, Illinois, 62701, USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given city.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with city: Springfield"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/stringify/region/{region}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve stringified locations filtered by region",
        "description": "This endpoint fetches all locations in a given region and returns them as formatted addresses.",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Region name to filter locations"
          },
          {
            "name": "city",
            "in": "path",
            "description": "City name to filter locations",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of stringified addresses for the specified region.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "123 Main St, Springfield, Illinois, 62701, USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given region.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with region: Illinois"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/stringify/country/{country}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve stringified locations filtered by country",
        "description": "This endpoint fetches all locations in a given country and returns them as formatted addresses.",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Country name to filter locations"
          },
          {
            "name": "city",
            "in": "path",
            "description": "City name to filter locations",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of stringified addresses for the specified country.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string",
                    "example": "123 Main St, Springfield, Illinois, 62701, USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given country.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with country: USA"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/stringify/{id}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve a stringified location filtered by ID",
        "description": "This endpoint fetches a location by ID and returns it as a formatted address.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Location ID to retrieve the location"
          },
          {
            "name": "city",
            "in": "path",
            "description": "City name to filter locations",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Stringified address for the specified location.",
            "schema": {
              "type": "object",
              "properties": {
                "address": {
                  "type": "string",
                  "example": "123 Main St, Apt 4B, Springfield, Illinois, 62701, USA"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No location found for the given ID.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with id: 1"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/city/{city}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve locations filtered by city",
        "description": "This endpoint fetches all locations in a given city.",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "City name to filter locations"
          }
        ],
        "responses": {
          "200": {
            "description": "List of locations in the specified city.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "streetAddress": {
                    "type": "string",
                    "example": "123 Main St"
                  },
                  "secondaryAddress": {
                    "type": "string",
                    "example": "Apt 4B"
                  },
                  "city": {
                    "type": "string",
                    "example": "Springfield"
                  },
                  "region": {
                    "type": "string",
                    "example": "Illinois"
                  },
                  "zipCode": {
                    "type": "string",
                    "example": "62701"
                  },
                  "country": {
                    "type": "string",
                    "example": "USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given city.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with city: Springfield"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/region/{region}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve locations filtered by region",
        "description": "This endpoint fetches all locations in a given region.",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Region name to filter locations"
          }
        ],
        "responses": {
          "200": {
            "description": "List of locations in the specified region.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "streetAddress": {
                    "type": "string",
                    "example": "123 Main St"
                  },
                  "secondaryAddress": {
                    "type": "string",
                    "example": "Apt 4B"
                  },
                  "city": {
                    "type": "string",
                    "example": "Springfield"
                  },
                  "region": {
                    "type": "string",
                    "example": "Illinois"
                  },
                  "zipCode": {
                    "type": "string",
                    "example": "62701"
                  },
                  "country": {
                    "type": "string",
                    "example": "USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given region.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with region: Illinois"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/location/country/{country}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Retrieve locations filtered by country",
        "description": "This endpoint fetches all locations in a given country.",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Country name to filter locations"
          }
        ],
        "responses": {
          "200": {
            "description": "List of locations in the specified country.",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "streetAddress": {
                    "type": "string",
                    "example": "123 Main St"
                  },
                  "secondaryAddress": {
                    "type": "string",
                    "example": "Apt 4B"
                  },
                  "city": {
                    "type": "string",
                    "example": "Springfield"
                  },
                  "region": {
                    "type": "string",
                    "example": "Illinois"
                  },
                  "zipCode": {
                    "type": "string",
                    "example": "62701"
                  },
                  "country": {
                    "type": "string",
                    "example": "USA"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "No locations found for the given country.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No such location with country: USA"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message describing the issue."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}